welcome
Created welcome page and added tests by inverting the render control.
Rather than "import { h } from 'hyperapp'" in all of the views, the
views now accept "h" as a function argument, allowing the test cases to
mock it. This also had some interesting benefits, such as not needing to
use babel when testing, and also being able to push side effects (such
as changes to the document head) to main.js.