remove elements.js
Removing elements.js since it was not really necessary. This could
eventually be useful again if we investigated using functional
composition to compose view components. This would also have the added
benefit of removing "h" from all of the views. Before doing this, we
should consider what kind of performance hits this will incur.